+++ index.js
@@ -1,6 +1,6 @@
-import DynamicClass from "./DynamicClass.js";
+import DynamicClass from "./src/DynamicClass.js";
 
 (() => {
   new DynamicClass("ClassOne");
-  new DynamicClass("ClassTwo", "Juna");
+  new DynamicClass("ClassTwo", "dave");
 })();
